# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ParkVision Smart Parking Platform API
  version: 1.0.0-draft
  description: >-
    Target REST contract for the ParkVision multi-tenant Smart Parking SaaS. This is a design
    artifact (Status: Draft, 2026-07-09) for issue DAI-260, aligned with docs/14_Backend_API.

    Tenancy is implicit: user requests resolve the tenant from the JWT (`tid` claim), while
    camera ingest resolves tenant/site from X-Camera-Id plus X-Camera-Key; PostgreSQL Row-Level
    Security enforces both paths. Legacy endpoints
    under `/api` (auth, vehicles/check-vehicle, gates, vehicle-logs, access-requests) remain and
    are documented in code via springdoc; this spec focuses on the new `/api/v1` SaaS surface plus
    the edge ingest contract.
  contact:
    name: Principal Architect
servers:
  - url: https://{tenant}.parkvision.example.com/api/v1
    description: Production (per-tenant subdomain)
    variables:
      tenant:
        default: acme
  - url: http://localhost:8080/api/v1
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /parking-events:
    get:
      tags: [Events]
      summary: Query the parking event log (timeline)
      parameters:
        - { name: siteId, in: query, schema: { type: string, format: uuid } }
        - { name: type, in: query, schema: { $ref: '#/components/schemas/EventType' } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ParkingEvent' }
    post:
      tags: [Ingest]
      summary: Authenticated, idempotent camera event ingest
      description: >-
        The camera identity is authenticated by X-Camera-Id and X-Camera-Key. Tenant and site are
        derived server-side from that credential. Send JSON, or multipart with an event JSON part
        and optional original_frame / plate_crop evidence parts. The legacy snapshot part remains
        accepted. Idempotency is scoped to (authenticated camera ID, event ID).
      security:
        - cameraId: []
          cameraKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/VehicleDetectedIngestEvent'
                - $ref: '#/components/schemas/PlateRecognizedIngestEvent'
                - $ref: '#/components/schemas/CameraIngestRequest'
          multipart/form-data:
            schema:
              type: object
              required: [event]
              properties:
                event: { $ref: '#/components/schemas/CameraIngestRequest' }
                snapshot:
                  type: string
                  format: binary
                  description: Legacy optional evidence image.
                original_frame:
                  type: string
                  format: binary
                  description: Optional original-frame evidence.
                plate_crop:
                  type: string
                  format: binary
                  description: Optional cropped-plate evidence. Combined evidence maximum is 5 MB.
            encoding:
              event:
                contentType: application/json
      responses:
        '202':
          description: Accepted or idempotent replay
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CameraIngestAccepted' }
        '400': { description: Malformed event or mismatched body camera ID }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: Snapshot exceeds the configured 5 MB limit }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cameraId:
      type: apiKey
      in: header
      name: X-Camera-Id
      description: Non-secret UUID of the camera that is presenting its ingest credential.
    cameraKey:
      type: apiKey
      in: header
      name: X-Camera-Key
      description: Per-camera ingest credential. Combined with X-Camera-Id, it derives tenant and site scope server-side.
...
  responses:
    Unauthorized:
      description: Missing/invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
...
    EventType:
      type: string
      enum:
        - MotionDetected
        - VehicleDetected
        - PlateRecognized
        - VehicleEntered
        - VehicleRelocated
        - VehicleExited
        - PersonDetected
        - SnapshotSaved
        - NotificationSent
    ParkingEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        siteId: { type: string, format: uuid }
        cameraId: { type: [string, 'null'], format: uuid }
        type: { $ref: '#/components/schemas/EventType' }
        licensePlate: { type: [string, 'null'] }
        trackId: { type: [string, 'null'] }
        slotId: { type: [string, 'null'], format: uuid }
        oldSlotId: { type: [string, 'null'], format: uuid }
        newSlotId: { type: [string, 'null'], format: uuid }
        personPresent: { type: [boolean, 'null'] }
        confidence: { type: [number, 'null'] }
        occurredAt: { type: string, format: date-time }
    CameraIngestRequest:
      type: object
      description: >-
        Current durable camera-ingest envelope. eventType is canonical; the runtime also accepts
        type as a legacy alias. cameraId is optional and, when supplied, must match X-Camera-Id.
        Tenant/site/object storage identity is derived by the backend and is never request input.
      required: [eventId, eventType, occurredAt]
      properties:
        eventId:
          type: string
          format: uuid
          description: Client-generated idempotency key, unique within the authenticated camera.
        eventType:
          type: string
          maxLength: 40
          description: Canonical event name; VehicleDetected and PlateRecognized have typed lpr-mvp-v1 profiles.
        cameraId:
          type: string
          format: uuid
          description: Optional echoed camera ID; cannot select or impersonate another camera.
        occurredAt:
          type: string
          format: date-time
          description: RFC 3339 detection time retained across offline retries.
        payload:
          type: object
          additionalProperties: true
          description: Generic structured event payload. Typed LPR profiles are additive conventions inside this field.
      additionalProperties: true
    VehicleDetectedIngestEvent:
      allOf:
        - $ref: '#/components/schemas/CameraIngestRequest'
        - type: object
          properties:
            eventType: { const: VehicleDetected }
            payload: { $ref: '#/components/schemas/VehicleDetectedPayload' }
          example:
            eventId: afc28cc1-7d60-4edb-a1d8-92881e00d8e4
            eventType: VehicleDetected
            cameraId: 6f5e50b3-b5a7-49b6-b341-bb16f1429100
            occurredAt: '2026-07-13T09:18:42.381+07:00'
            payload:
              eventVersion: 1
              pipeline: { id: lpr-mvp-v1, configurationHash: 'sha256:6e4b1d...' }
              frame: { capturedAt: '2026-07-13T09:18:42.381+07:00', width: 1920, height: 1080, coordinateSpace: original-frame-pixels }
              tracker: { sessionId: 8dc0d470-f897-4c15-abf9-03c7a8f0ac7f, trackId: '42' }
              vehicle:
                class: car
                confidence: 0.93
                boundingBox: { x: 492, y: 243, width: 916, height: 540 }
              models:
                vehicleDetector: { name: yolo11n, artifactVersion: '2026.07.0', confidenceThreshold: 0.4 }
    PlateRecognizedIngestEvent:
      allOf:
        - $ref: '#/components/schemas/CameraIngestRequest'
        - type: object
          properties:
            eventType: { const: PlateRecognized }
            payload: { $ref: '#/components/schemas/PlateRecognizedPayload' }
          example:
            eventId: 28e2e297-3285-45a3-89e5-94cc2e1fbbd3
            eventType: PlateRecognized
            cameraId: 6f5e50b3-b5a7-49b6-b341-bb16f1429100
            occurredAt: '2026-07-13T09:18:42.581+07:00'
            payload:
              eventVersion: 1
              causationEventId: afc28cc1-7d60-4edb-a1d8-92881e00d8e4
              pipeline: { id: lpr-mvp-v1, configurationHash: 'sha256:6e4b1d...' }
              frame: { capturedAt: '2026-07-13T09:18:42.581+07:00', width: 1920, height: 1080, coordinateSpace: original-frame-pixels }
              tracker: { sessionId: 8dc0d470-f897-4c15-abf9-03c7a8f0ac7f, trackId: '42' }
              vehicle:
                class: car
                confidence: 0.94
                boundingBox: { x: 492, y: 243, width: 916, height: 540 }
              plate:
                text: 51A-123.45
                normalizedText: 51A12345
                detectionConfidence: 0.94
                recognitionConfidence: 0.96
                boundingBox: { x: 780, y: 554, width: 242, height: 78 }
              snapshots:
                - { kind: original_frame, capturedAt: '2026-07-13T09:18:42.581+07:00', contentType: image/jpeg, width: 1920, height: 1080, sha256: 'sha256:0ad1...' }
                - { kind: plate_crop, contentType: image/jpeg, width: 242, height: 78, sha256: 'sha256:c62e...' }
              snapshotUpload: { part: snapshot, kind: plate_crop }
              models:
                vehicleDetector: { name: yolo11n, artifactVersion: '2026.07.0', confidenceThreshold: 0.4 }
                plateDetector: { name: lp-detector-nano, artifactVersion: '61', confidenceThreshold: 0.6 }
                ocr: { name: PaddleOCR, artifactVersion: pp-ocr-mobile, recognitionConfidenceThreshold: 0.8 }
    VehicleDetectedPayload:
      type: object
      required: [eventVersion, pipeline, frame, tracker, vehicle, models]
      properties:
        eventVersion: { const: 1 }
        pipeline: { $ref: '#/components/schemas/PipelineProvenance' }
        frame: { $ref: '#/components/schemas/FrameMetadata' }
        tracker: { $ref: '#/components/schemas/TrackerIdentity' }
        vehicle: { $ref: '#/components/schemas/VehicleDetection' }
        models:
          type: object
          required: [vehicleDetector]
          properties:
            vehicleDetector: { $ref: '#/components/schemas/ModelProvenance' }
        snapshots:
          type: array
          description: Descriptive evidence metadata; enter events may include the original frame.
          items: { $ref: '#/components/schemas/SnapshotMetadata' }
        snapshotUpload:
          type: object
          description: "When present on VehicleDetected, the optional multipart snapshot is the original frame."
          properties:
            part: { const: snapshot }
            kind: { type: string, enum: [original_frame, plate_crop] }
    PlateRecognizedPayload:
      allOf:
        - $ref: '#/components/schemas/VehicleDetectedPayload'
        - type: object
          required: [causationEventId, plate, models]
          properties:
            causationEventId: { type: string, format: uuid }
            plate: { $ref: '#/components/schemas/PlateRecognition' }
            snapshots:
              type: array
              description: Descriptive evidence metadata only; object keys and URLs are server-generated.
              items: { $ref: '#/components/schemas/SnapshotMetadata' }
            snapshotUpload:
              type: object
              description: "Current multipart convention: the one optional snapshot part is a plate crop."
              properties:
                part: { const: snapshot }
                kind: { const: plate_crop }
            models:
              type: object
              required: [vehicleDetector, plateDetector, ocr]
              properties:
                vehicleDetector: { $ref: '#/components/schemas/ModelProvenance' }
                plateDetector: { $ref: '#/components/schemas/ModelProvenance' }
                ocr: { $ref: '#/components/schemas/ModelProvenance' }
    PipelineProvenance:
      type: object
      required: [id, configurationHash]
      properties:
        id: { const: lpr-mvp-v1 }
        configurationHash: { type: string, pattern: '^sha256:' }
    FrameMetadata:
      type: object
      required: [capturedAt, width, height, coordinateSpace]
      properties:
        capturedAt: { type: string, format: date-time }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        coordinateSpace: { const: original-frame-pixels }
    TrackerIdentity:
      type: object
      required: [sessionId, trackId]
      properties:
        sessionId: { type: string, format: uuid }
        trackId: { type: string, minLength: 1 }
    BoundingBox:
      type: object
      required: [x, y, width, height]
      properties:
        x: { type: integer, minimum: 0 }
        y: { type: integer, minimum: 0 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
    VehicleDetection:
      type: object
      required: [class, confidence, boundingBox]
      properties:
        class: { type: string, enum: [car, motorbike] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        boundingBox: { $ref: '#/components/schemas/BoundingBox' }
    PlateRecognition:
      type: object
      required: [text, normalizedText, detectionConfidence, recognitionConfidence, boundingBox]
      properties:
        text: { type: string, minLength: 1 }
        normalizedText: { type: string, pattern: '^[A-Z0-9]+$' }
        detectionConfidence: { type: number, minimum: 0, maximum: 1 }
        recognitionConfidence: { type: number, minimum: 0, maximum: 1 }
        boundingBox: { $ref: '#/components/schemas/BoundingBox' }
    ModelProvenance:
      type: object
      required: [name, artifactVersion]
      properties:
        name: { type: string }
        artifactVersion: { type: string }
        confidenceThreshold: { type: number, minimum: 0, maximum: 1 }
        recognitionConfidenceThreshold: { type: number, minimum: 0, maximum: 1 }
    SnapshotMetadata:
      type: object
      required: [kind, contentType, width, height, sha256]
      properties:
        kind: { type: string, enum: [original_frame, plate_crop] }
        capturedAt: { type: string, format: date-time }
        contentType: { type: string, pattern: '^image/' }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        sha256: { type: string, pattern: '^sha256:' }
        sourceBoundingBox: { $ref: '#/components/schemas/BoundingBox' }
    CameraIngestAccepted:
      type: object
      required: [eventId, status]
      properties:
        eventId: { type: string, format: uuid }
        status: { const: accepted }
...